# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Título da API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Lista os produtos",
                "description": "Retorna uma lista paginada de produtos",
                "operationId": "listProducts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para paginação",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{code}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Detalhes de um produto",
                "operationId": "aa1c2d7e05e44e82699d584503d190ea",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ABC123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "ABC123"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Arroz Tio João 5kg"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 25.99
                                                },
                                                "stock": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Deleta um produto",
                "operationId": "50cfcc9c8fef302921bf0613c13c544b",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ABC123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produto deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{code}": {
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualiza um produto",
                "description": "Atualiza um produto existente com os dados enviados.",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "status",
                                    "imported_t",
                                    "url",
                                    "creator",
                                    "created_t",
                                    "last_modified_t",
                                    "product_name"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer",
                                        "example": 20221126
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "published",
                                            "draft",
                                            "trash"
                                        ],
                                        "example": "published"
                                    },
                                    "imported_t": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2020-02-07T16:00:00Z"
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://world.openfoodfacts.org/product/20221126"
                                    },
                                    "creator": {
                                        "type": "string",
                                        "example": "securita"
                                    },
                                    "created_t": {
                                        "type": "integer",
                                        "example": 1415302075
                                    },
                                    "last_modified_t": {
                                        "type": "integer",
                                        "example": 1572265837
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Madalenas quadradas"
                                    },
                                    "quantity": {
                                        "type": "string",
                                        "example": "380 g (6 x 2 u.)"
                                    },
                                    "brands": {
                                        "type": "string",
                                        "example": "La Cestera"
                                    },
                                    "categories": {
                                        "type": "string",
                                        "example": "Bolos, Lanches doces"
                                    },
                                    "labels": {
                                        "type": "string",
                                        "example": "Contem gluten, Contém ovos"
                                    },
                                    "cities": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "purchase_places": {
                                        "type": "string",
                                        "example": "Braga,Portugal"
                                    },
                                    "stores": {
                                        "type": "string",
                                        "example": "Lidl"
                                    },
                                    "ingredients_text": {
                                        "type": "string",
                                        "example": "farinha de trigo, açúcar, óleo vegetal"
                                    },
                                    "traces": {
                                        "type": "string",
                                        "example": "Leite, Soja"
                                    },
                                    "serving_size": {
                                        "type": "string",
                                        "example": "madalena 31.7 g"
                                    },
                                    "serving_quantity": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 31.7
                                    },
                                    "nutriscore_score": {
                                        "type": "integer",
                                        "example": 17
                                    },
                                    "nutriscore_grade": {
                                        "type": "string",
                                        "enum": [
                                            "a",
                                            "b",
                                            "c",
                                            "d",
                                            "e"
                                        ],
                                        "example": "d"
                                    },
                                    "main_category": {
                                        "type": "string",
                                        "example": "en:madeleines"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://static.openfoodfacts.org/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 20221126
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "published"
                                        },
                                        "product_name": {
                                            "type": "string",
                                            "example": "Madalenas quadradas"
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "format": "url",
                                            "example": "https://static.openfoodfacts.org/image.jpg"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-22T15:03:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Produto",
                "required": [
                    "id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Arroz Branco"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.9
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-22T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-22T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}